change function for stockController from RequestStockController also update web
